split helper functions part 9
