+++ .vscode/settings.json
@@ -1,0 +1,5 @@
+{
+    "Lua.diagnostics.disable": [
+        "lowercase-global"
+    ]
+}
